Move initial analysis to subfolder
Initially, this repository contained a single Rmd analysis file
containing analysis of the 37d experiment done by Dan and Ari. This
commit moves that Rmd and associated files to a subfolder. The data
folder is kept at the root level; a symlink is added to maintain valid
paths to data files in the Rmd file.